Fix Customer ID bug
Summary:
- Apparently customer ID was never autogenerating, since we had UUID as the id column
fixed that
- Also expose a DELETE endpoint for customer

Test Plan: sbt test

Reviewers: bharat, pritam, karthik

Reviewed By: karthik

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D554

+++ managed/src/main/java/com/yugabyte/yw/models/Universe.java
@@ -60,7 +60,7 @@ public class Universe extends Model {
 
   // The customer id, needed only to enforce unique universe names for a customer.
   @Constraints.Required
-  public int customerId;
+  public Long customerId;
 
   // The Json serialized version of universeDetails. This is used only in read from and writing to
   // the DB.
@@ -110,7 +110,7 @@ public class Universe extends Model {
    * @param name : name of the universe
    * @return the newly created universe
    */
-  public static Universe create(String name, int customerId) {
+  public static Universe create(String name, Long customerId) {
     // Create the universe object.
     Universe universe = new Universe();
